perezaliv2

+++ VKAPI/Handlers/Board.php
@@ -22,6 +22,10 @@ final class Board extends VKAPIRequestHandler
             $this->fail(403, "Invalid club");
         }
 
+        if($club->isDeleted()) {
+            $this->fail(43, "Club was deleted");
+        }
+
         if(!$club->canBeModifiedBy($this->getUser()) && !$club->isEveryoneCanCreateTopics()) {
             $this->fail(403, "Access to club denied");
         }
@@ -111,6 +115,10 @@ final class Board extends VKAPIRequestHandler
             return 0;
         }
 
+        if($topic->getClub()->isDeleted()) {
+            $this->fail(43, "Club was deleted");
+        }
+
         if(!$topic->isClosed()) {
             $topic->setClosed(1);
             $topic->save();
@@ -130,8 +138,16 @@ final class Board extends VKAPIRequestHandler
 
         $topic = (new TopicsRepo)->getTopicById($group_id, $topic_id);
 
-        if(!$topic || $topic->isDeleted() || $topic->isClosed()) {
-            $this->fail(100, "Topic is deleted, closed or invalid.");
+        if(!$topic || $topic->isClosed()) {
+            $this->fail(100, "Topic is deleted or invalid.");
+        }
+
+        if(!$topic->getClub() || $topic->getClub()->isDeleted()) {
+            $this->fail(43, "Club was deleted");
+        }
+
+        if(!$topic->canBeViewedBy($this->getUser())) {
+            $this->fail(8, "Access denied");
         }
 
         $flags = 0;
@@ -224,6 +240,10 @@ final class Board extends VKAPIRequestHandler
             return 0;
         }
 
+        if($topic->getClub()->isDeleted()) {
+            return 0;
+        }
+
         $topic->deleteTopic();
 
         return 1;
@@ -258,6 +278,10 @@ final class Board extends VKAPIRequestHandler
             return 0;
         }
 
+        if($topic->getClub()->isDeleted()) {
+            return 0;
+        }
+
         $topic->setTitle(ovk_proc_strtr($title, 127));
 
         $topic->save();
@@ -273,6 +297,10 @@ final class Board extends VKAPIRequestHandler
         $topic = (new TopicsRepo)->getTopicById($group_id, $topic_id);
 
         if(!$topic || !$topic->getClub() || !$topic->getClub()->canBeModifiedBy($this->getUser())) {
+            return 0;
+        }
+
+        if($topic->getClub()->isDeleted()) {
             return 0;
         }
 
@@ -293,6 +321,14 @@ final class Board extends VKAPIRequestHandler
 
         if(!$topic || !$topic->getClub() || $topic->isDeleted()) {
             $this->fail(5, "Invalid topic");
+        }
+
+        if(!$topic->canBeViewedBy($this->getUser())) {
+            $this->fail(8, "Access denied");
+        }
+
+        if(!$topic->getClub() || $topic->getClub()->isDeleted()) {
+            $this->fail(666, "Club was deleted or banned");
         }
 
         $arr = [
@@ -325,6 +361,14 @@ final class Board extends VKAPIRequestHandler
 
         $arr = [];
         $club = (new ClubsRepo)->get($group_id);
+
+        if(!$club || $club->isDeleted()) {
+            $this->fail(666, "Club was deleted or banned");
+        }
+
+        if(!$club->canBeViewedBy($this->getUser())) {
+            $this->fail(8, "Access denied");
+        }
 
         $topics = array_slice(iterator_to_array((new TopicsRepo)->getClubTopics($club, 1, $count + $offset)), $offset);
         $arr["count"] = (new TopicsRepo)->getClubTopicsCount($club);
@@ -365,6 +409,10 @@ final class Board extends VKAPIRequestHandler
             return 0;
         }
 
+        if(!$topic->getClub() || $topic->getClub()->isDeleted()) {
+            $this->fail(666, "Club was deleted or banned");
+        }
+
         if($topic->isClosed()) {
             $topic->setClosed(0);
             $topic->save();
@@ -387,6 +435,10 @@ final class Board extends VKAPIRequestHandler
 
         if(!$topic || !$topic->getClub() || !$topic->getClub()->canBeModifiedBy($this->getUser())) {
             return 0;
+        }
+
+        if(!$topic->getClub() || $topic->getClub()->isDeleted()) {
+            $this->fail(666, "Club was deleted or banned");
         }
 
         if($topic->isPinned()) {
